aux_tasks/requirements.txt
absl-py==1.1.0
ale-py==0.7.5
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
chex==0.1.3
cloudpickle==2.1.0
clu==0.0.7
contextlib2==21.6.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.5.1
dm-reverb==0.7.3
dm-tree==0.1.7
dopamine-rl==4.0.5
etils==0.6.0
flatbuffers==2.0
flax==0.5.0
fonttools==4.33.3
frozendict==2.3.2
gast==0.5.3
gin-config==0.5.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
grpcio==1.46.3
gym==0.24.0
gym-minigrid==1.0.3
gym-notices==0.0.6
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
jax==0.3.13
jaxlib==0.3.10
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
ml-collections==0.1.1
msgpack==1.0.4
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
portpicker==1.5.0
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-datasets==4.6.0
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
tensorflow-probability==0.16.0
termcolor==1.1.0
tf-slim==1.1.0
toml==0.10.2
toolz==0.11.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
